requirements.txt
requests
sqlalchemy
psycopg2-binary
pyTelegramBotAPI
python-dotenv
